refactor(server): extract compaction scheduler startup to core/compaction_scheduler.py
Task F of docs/current/server-py-module-split.md. Pure extraction,
verbatim move verified byte-identical against git history
(_maybe_start_compaction_scheduler, 38 lines). Removed now-dead
COMPACTION_AUTO_APPLY_ENABLED/COMPACTION_AUTO_APPLY_INTERVAL_MINUTES/
COMPACTION_ENABLED imports from server.py (SCHEDULER_AVAILABLE stays,
still used by check_dependencies and main()'s feature-status log).

The function's two inline lazy imports needed their relative-import
depth updated for the new package level (core/ is one level below
marm_mcp_server/, same as services/ and endpoints/, so `.services...`
and `.endpoints...` become `..services...` and `..endpoints...`) --
this is the one narrow, spec-mandated wiring fix inside the moved
body; diffed against the verbatim slice to confirm those are the only
two lines touched. lifespan()'s call site is unchanged.

test_compaction_auto_apply.py had 3 tests using
monkeypatch.setattr(server_mod, "COMPACTION_AUTO_APPLY_ENABLED", ...)
against marm_mcp_server.server -- same category of issue as Task E:
the function now resolves these names as globals in
core/compaction_scheduler.py's own module dict, so the monkeypatch
target had to move there too for the patch to actually reach the
function. Repointed to marm_mcp_server.core.compaction_scheduler; no
assertion or test logic changed.

Functional check runs the real extracted function under a live asyncio
event loop (APScheduler requires a running loop): confirms it starts
a genuine AsyncIOScheduler with the compaction_auto_apply job
registered on the correct interval trigger for both the
auto-apply-enabled and auto-apply-disabled code paths, and confirms
the COMPACTION_ENABLED=False path still returns None. 76 targeted
tests pass across test_compaction_staging.py,
test_compaction_auto_apply.py, and test_cli_entrypoint.py.
